+++ centaur.cpp
@@ -2,12 +2,12 @@
 
 Centaur::Centaur() : Vehicle() {
 	name = "Кентавр";
-	speed = 15;
+	speed = 15.0;
 	moveTime = 8;
 }
 
 void Centaur::racing(int distance) {
-	result = distance / speed;
+	result = distance / double(speed);
 	int stops = result / moveTime;
 	double stopsTime = stops * 2;
